Tabla de materiales imputados

+++ Proyecto FUXA/Models/ImputacionMaterial.cs
@@ -8,7 +8,7 @@ public class ImputacionMaterial
     public int Id { get; set; }
     public int IdOperacion { get; set; }
     public int IdMaterial { get; set; }
-    public double Cantidad { get; set; }
+    public decimal Cantidad { get; set; }
     public string? Observaciones { get; set; }
     public DateTime FechaRegistro { get; set; }
